Working Info button and Center display templet done (#33)
* Added Info button with pop up - needs data population

Need to fix alignment of "Today" and add stuff to the pop up

* Working Button with pop up - Version 1

working button and pop up version 1. Im going to try to make the pop up it look nicer.

* Fixed the pop to be smaller, and not as intrusive

Added a "Today" text and fixed the pop up to be 40% of the screen instead of the whole screen

* WORKING VERSION WITH POP UP AND 3 MIDDLE BUTTONS

added 3 center buttons. Still Need to populate them with Images

* Added 3 buttons with icons

Added the 3 home view buttons with selectable image on the left for plant choice

* Images for Icons on center button